backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pyyaml==6.0.2
mistune==3.0.2
minio==7.2.12
weasyprint==62.3
mailjet-rest==1.3.4
slowapi==0.1.9
authlib==1.4.0
httpx==0.28.1
clamd==1.0.2
